Completed nowplaying ext, break screen and sponsors
The nowplaying extension is completed. It is intended to be
used with the music player Foobar2k and should watch the output
of the file showing the the currently playing song. It also has its
own dahboard panel.

The break screen is completed, with additional room made for the
nowplaying item.

A small tweak to the design was made where most border corners are now
rounded. Sponsor rotations have been added to each game overlay.